pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "pygeostat"
dynamic = ["version"]
authors = [
    {name = "Jared Deutsch"},
    {name = "Matthew Deutsch"},
    {name = "Ryan Martin"},
    {name = "Warren Black"},
    {name = "Tyler Acorn"},
    {name = "Mostafa Hadavand"},
    {name = "Ryan Barnett"},
]
maintainers = [
    {name = "Ronny A. Hernandez Mora", email = "ronnyale@ualberta.ca"},
]
description = "Python interface for geostatistics"
readme = "README.md"
keywords = ["geostatistics"]
license = {text = "MIT"}
classifiers = [
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.10"
dependencies = [
    "h5py>=3.13.0",
    "matplotlib>=3.10.0",
    "numba>=0.61.0",
    "numpy>=2.1.3",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "probscale>=0.2.5",
    "pyperclip>=1.9.0",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.2",
    "statsmodels>=0.14.4",
    "tqdm>=4.67.1",
]

[project.optional-dependencies]
dev = [
    "black>=25.1.0",
    "jupyter>=1.1.1",
    "jupyter-client>=8.6.3",
    "mypy>=1.15.0",
    "myst-parser>=4.0.1",
    "nbconvert>=7.16.6",
    "nbformat>=5.10.4",
    "pyevtk>=1.6.0",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.6",
    "sphinx>=8.1.3",
    "sphinx-rtd-theme>=3.0.2",
]
docs = [
    "sphinx>=8.1.3",
    "nbsphinx>=0.8.8",
    "sphinx-rtd-theme>=3.0.2",
    "matplotlib>=3.10.0",  
    "ipython>=8.0.0",      
    "myst-parser>=4.0.1" 
]

[project.urls]
Documentation = "https://www.ccgalberta.com/pygeostat/welcome.html"
Source = "https://github.com/CcgAlberta/pygeostat"
Issues = "https://github.com/CcgAlberta/pygeostat/issues"

[tool.setuptools.packages.find]
where = ["src"]
include = [
    "pygeostat",
    "pygeostat.*",
]

[tool.setuptools.dynamic]
version = {attr = "pygeostat.__version__"}

